requirements.txt
tensorboardx==2.4.1
gputil==1.4.0
numpy==1.20.3
pytorch==1.10.1
tqdm==4.62.3
matplotlib==3.4.3
requests==2.27.1
chardet==4.0.0
scipy==1.8.0